+++ README.md
@@ -2,26 +2,26 @@
 
 Please refer to [this Ethereum Research post](https://ethresear.ch/t/efficient-ecdsa-signature-verification-using-circom/13629) for details. The circuits in this repo uses circuits from [circom-ecdsa](https://github.com/0xPARC/circom-ecdsa).
 
-## Install dependencies
+### Install dependencies
 
 ```
 yarn
 ```
 
-## Run proof generation
+### Compile the circuit and generate the zkey
 
 ```
-yarn run prove
+yarn run build:ecdsaverify
+```
+
+### Run proof generation
+
+```
+yarn run run:ecdsaverify
 ```
 
 ## Benchmarks
 
-On a MacBook Pro
-
-_Full proving time includes public key -> eth address conversion._
-| Full proving time | 23s    |
-| ----------------- | ------ |
-| Proving key size  | 221MB  |
-| Constraints       | 401319 |
-
-_When the proving is done in a browser, it might take some time to download the proving key, which is 256MB in size._
+| Circuit     | Constraints | Full proving time <br /> (M1 Pro MacBook Pro) | zKey size |
+| ----------- | ----------- | --------------------------------------------- | --------- |
+| ecdsaverify | 163,239     | 24s                                           | 119MB     |
