+++ scripts/utils/point-cache.js
